[AbstractModeConsumer] Migrate order management methods
Creation of order_adapter.py
Creation of order_util.py
