+++ octobot_trading/consumers/abstract_mode_consumer.py
@@ -13,8 +13,6 @@
 #
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
-import asyncio
-
 from octobot_commons.logging.logging_util import get_logger
 from octobot_commons.symbol_util import split_symbol
 
@@ -25,30 +23,18 @@ from octobot_trading.enums import ExchangeConstantsMarketStatusColumns as Ecmsc,
 class AbstractTradingModeConsumer(ExchangeChannelInternalConsumer):
     def __init__(self, trading_mode):
         super().__init__()
-        self.logger = get_logger(self.__class__.__name__)
+        self._logger = get_logger(self.__class__.__name__)
 
         self.trading_mode = trading_mode
         self.exchange_manager = trading_mode.exchange_manager
-        self.trader = self.exchange_manager.trader
 
-        # shortcuts
-        self.exchange_personal_data = self.exchange_manager.exchange_personal_data
-        self.exchange_portfolio_manager = self.exchange_personal_data.portfolio_manager
-
-    # @abstractmethod
-    # async def internal_callback(self, **kwargs):
-    #     raise NotImplementedError("internal_callback is not implemented")
-
-    async def get_holdings_ratio(self, currency):
-        return await self.exchange_portfolio_manager.portfolio_profitability.holdings_ratio(currency)
-
-    def get_number_of_traded_assets(self):
-        return len(self.exchange_portfolio_manager.portfolio_profitability.origin_crypto_currencies_values)
+    async def internal_callback(self, **kwargs):
+        raise NotImplementedError("internal_callback is not implemented")
 
     # Can be overwritten
     async def can_create_order(self, symbol, state):
         currency, market = split_symbol(symbol)
-        portfolio = self.exchange_portfolio_manager
+        portfolio = self.exchange_manager.exchange_personal_data.portfolio_manager
 
         # get symbol min amount when creating order
         symbol_limit = self.exchange_manager.exchange.get_market_status(symbol)[Ecmsc.LIMITS.value]
@@ -69,24 +55,28 @@ class AbstractTradingModeConsumer(ExchangeChannelInternalConsumer):
         # other cases like neutral state or unfulfilled previous conditions
         return False
 
-    async def get_pre_order_data(self, symbol):
-        try:
-            mark_price = await self.exchange_manager.exchange_symbols_data.get_exchange_symbol_data(symbol)\
-                    .prices_manager.get_mark_price()
-        except asyncio.TimeoutError:
-            raise ValueError(f"Mark price is not available")
+    @staticmethod
+    def check_factor(min_val, max_val, factor):
+        """
+        Checks if factor is min_val < factor < max_val
+        :param min_val:
+        :param max_val:
+        :param factor:
+        :return:
+        """
+        if factor > max_val:
+            return max_val
+        if factor < min_val:
+            return min_val
+        return factor
 
-        currency, market = split_symbol(symbol)
+    async def get_holdings_ratio(self, currency):
+        return await self.exchange_manager.exchange_personal_data.portfolio_manager.portfolio_profitability \
+            .holdings_ratio(currency)
 
-        current_symbol_holding = self.exchange_portfolio_manager.portfolio.get_currency_portfolio(currency)
-        current_market_quantity = self.exchange_portfolio_manager.portfolio.get_currency_portfolio(market)
-
-        market_quantity = current_market_quantity / mark_price
-
-        symbol_market = self.exchange_manager.exchange.get_market_status(symbol, with_fixer=False)
-
-        return current_symbol_holding, current_market_quantity, market_quantity, mark_price, symbol_market
-
+    def get_number_of_traded_assets(self):
+        return len(self.exchange_manager.exchange_personal_data.portfolio_manager.portfolio_profitability
+                   .origin_crypto_currencies_values)
 
 # class AbstractTradingModeConsumerWithBot(AbstractTradingModeConsumer, Initializable):  # TODO
 #     def __init__(self, trading_mode, sub_portfolio_percent):
