+++ octobot_trading/orders/order_adapter.py
@@ -14,12 +14,13 @@
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
 import math
-from octobot_commons.dict_util import get_value_or_default
-from octobot_commons.logging.logging_util import get_logger
 
+from octobot_commons.dict_util import get_value_or_default
 from octobot_trading.constants import CURRENCY_DEFAULT_MAX_PRICE_DIGITS
+
 from octobot_trading.enums import ExchangeConstantsMarketStatusColumns as Ecmsc
-from octobot_trading.exchanges.util.exchange_market_status_fixer import ExchangeMarketStatusFixer, is_ms_valid
+from octobot_trading.exchanges.util.exchange_market_status_fixer import ExchangeMarketStatusFixer
+from octobot_trading.orders.order_util import is_valid, check_cost
 
 
 def adapt_price(symbol_market, price):
@@ -35,7 +36,7 @@ def adapt_quantity(symbol_market, quantity):
     return trunc_with_n_decimal_digits(quantity, maximal_volume_digits)
 
 
-def trunc_with_n_decimal_digits(value, digits):
+def trunc_with_n_decimal_digits(value, digits):  # TODO migrate to commons
     # force exact representation
     return float("{0:.{1}f}".format(math.trunc(value * 10 ** digits) / (10 ** digits), digits if digits > 1 else 1))
 
@@ -71,90 +72,38 @@ def adapt_order_quantity_because_price(limiting_value, max_value, price, symbol_
     return orders
 
 
-def check_factor(min_val, max_val, factor):
+def split_orders(total_order_price, max_cost, valid_quantity, max_quantity, price, quantity, symbol_market):
     """
-    Checks if factor is min_val < factor < max_val
-    :param min_val:
-    :param max_val:
-    :param factor:
-    :return:
-    """
-    if factor > max_val:
-        return max_val
-    if factor < min_val:
-        return min_val
-    return factor
-
-
-def is_valid(element, key):
-    """
-    Checks is the element is valid with the market status fixer
-    :param element:
-    :param key:
-    :return:
-    """
-    return key in element and is_ms_valid(element[key])
-
-
-def get_min_max_amounts(symbol_market, default_value=None):
-    """
-    Returns the min and max quantity, cost and price according to the specified market
+    Splits too big orders into multiple ones according to the max_cost and max_quantity
+    :param total_order_price:
+    :param max_cost:
+    :param valid_quantity:
+    :param max_quantity:
+    :param price:
+    :param quantity:
     :param symbol_market:
-    :param default_value:
     :return:
     """
-    min_quantity = max_quantity = min_cost = max_cost = min_price = max_price = default_value
-    if Ecmsc.LIMITS.value in symbol_market:
-        symbol_market_limits = symbol_market[Ecmsc.LIMITS.value]
+    if max_cost is None and max_quantity is None:
+        raise RuntimeError("Impossible to split orders with max_cost and max_quantity undefined.")
+    nb_orders_according_to_cost = None
+    nb_orders_according_to_quantity = None
+    if max_cost:
+        nb_orders_according_to_cost = total_order_price / max_cost
 
-        if Ecmsc.LIMITS_AMOUNT.value in symbol_market_limits:
-            limit_amount = symbol_market_limits[Ecmsc.LIMITS_AMOUNT.value]
-            if is_valid(limit_amount, Ecmsc.LIMITS_AMOUNT_MIN.value) \
-                    or is_valid(limit_amount, Ecmsc.LIMITS_AMOUNT_MAX.value):
-                min_quantity = get_value_or_default(limit_amount, Ecmsc.LIMITS_AMOUNT_MIN.value, default_value)
-                max_quantity = get_value_or_default(limit_amount, Ecmsc.LIMITS_AMOUNT_MAX.value, default_value)
+    if max_quantity:
+        nb_orders_according_to_quantity = valid_quantity / max_quantity
 
-        # case 2: use cost and price
-        if Ecmsc.LIMITS_COST.value in symbol_market_limits:
-            limit_cost = symbol_market_limits[Ecmsc.LIMITS_COST.value]
-            if is_valid(limit_cost, Ecmsc.LIMITS_COST_MIN.value) \
-                    or is_valid(limit_cost, Ecmsc.LIMITS_COST_MAX.value):
-                min_cost = get_value_or_default(limit_cost, Ecmsc.LIMITS_COST_MIN.value, default_value)
-                max_cost = get_value_or_default(limit_cost, Ecmsc.LIMITS_COST_MAX.value, default_value)
-
-        # case 2: use quantity and price
-        if Ecmsc.LIMITS_PRICE.value in symbol_market_limits:
-            limit_price = symbol_market_limits[Ecmsc.LIMITS_PRICE.value]
-            if is_valid(limit_price, Ecmsc.LIMITS_PRICE_MIN.value) \
-                    or is_valid(limit_price, Ecmsc.LIMITS_PRICE_MAX.value):
-                min_price = get_value_or_default(limit_price, Ecmsc.LIMITS_PRICE_MIN.value, default_value)
-                max_price = get_value_or_default(limit_price, Ecmsc.LIMITS_PRICE_MAX.value, default_value)
-
-    return min_quantity, max_quantity, min_cost, max_cost, min_price, max_price
-
-
-def check_cost(total_order_price, min_cost):
-    """
-    Checks and adapts the quantity and price of the order to ensure it's exchange compliant:
-    - are the quantity and price of the order compliant with the exchange's number of digits requirement
-        => otherwise quantity will be truncated accordingly
-    - is the quantity valid
-    - are the order total price and quantity superior or equal to the exchange's minimum order requirement
-        => otherwise order is impossible => returns empty list
-    - if total cost data are unavailable:
-    - is the price of the currency compliant with the exchange's price interval for this currency
-        => otherwise order is impossible => returns empty list
-    - are the order total price and quantity inferior or equal to the exchange's maximum order requirement
-        => otherwise order is impossible as is => split order into smaller ones and returns the list
-    => returns the quantity and price list of possible order(s)
-    - if exchange symbol data are not enough
-        => try fixing exchange data using ExchangeMarketStatusFixer are start again (once only)
-    """
-    if total_order_price < min_cost:
-        if min_cost is None:
-            get_logger().error("Invalid min_cost from exchange")
-        return False
-    return True
+    if nb_orders_according_to_cost is None:
+        # can only split using quantity
+        return adapt_order_quantity_because_quantity(valid_quantity, max_quantity, quantity, price, symbol_market)
+    elif nb_orders_according_to_quantity is None:
+        # can only split using price
+        return adapt_order_quantity_because_price(total_order_price, max_cost, price, symbol_market)
+    else:
+        if nb_orders_according_to_cost > nb_orders_according_to_quantity:
+            return adapt_order_quantity_because_price(total_order_price, max_cost, price, symbol_market)
+        return adapt_order_quantity_because_quantity(valid_quantity, max_quantity, quantity, price, symbol_market)
 
 
 def check_and_adapt_order_details_if_necessary(quantity, price, symbol_market, fixed_symbol_data=False):
@@ -244,40 +193,6 @@ def check_and_adapt_order_details_if_necessary(quantity, price, symbol_market, f
         return []
 
 
-def split_orders(total_order_price, max_cost, valid_quantity, max_quantity, price, quantity, symbol_market):
-    """
-    Splits too big orders into multiple ones according to the max_cost and max_quantity
-    :param total_order_price:
-    :param max_cost:
-    :param valid_quantity:
-    :param max_quantity:
-    :param price:
-    :param quantity:
-    :param symbol_market:
-    :return:
-    """
-    if max_cost is None and max_quantity is None:
-        raise RuntimeError("Impossible to split orders with max_cost and max_quantity undefined.")
-    nb_orders_according_to_cost = None
-    nb_orders_according_to_quantity = None
-    if max_cost:
-        nb_orders_according_to_cost = total_order_price / max_cost
-
-    if max_quantity:
-        nb_orders_according_to_quantity = valid_quantity / max_quantity
-
-    if nb_orders_according_to_cost is None:
-        # can only split using quantity
-        return adapt_order_quantity_because_quantity(valid_quantity, max_quantity, quantity, price, symbol_market)
-    elif nb_orders_according_to_quantity is None:
-        # can only split using price
-        return adapt_order_quantity_because_price(total_order_price, max_cost, price, symbol_market)
-    else:
-        if nb_orders_according_to_cost > nb_orders_according_to_quantity:
-            return adapt_order_quantity_because_price(total_order_price, max_cost, price, symbol_market)
-        return adapt_order_quantity_because_quantity(valid_quantity, max_quantity, quantity, price, symbol_market)
-
-
 def add_dusts_to_quantity_if_necessary(quantity, price, symbol_market, current_symbol_holding):
     """
     Adds remaining quantity to the order if the remaining quantity is too small
@@ -298,7 +213,7 @@ def add_dusts_to_quantity_if_necessary(quantity, price, symbol_market, current_s
 
     if not (is_valid(limit_amount, Ecmsc.LIMITS_AMOUNT_MIN.value) and
             is_valid(limit_cost, Ecmsc.LIMITS_COST_MIN.value)):
-        fixed_market_status = ExchangeMarketStatusFixer(symbol_market, price).get_market_status()
+        fixed_market_status = ExchangeMarketStatusFixer(symbol_market, price).market_status
         limit_amount = fixed_market_status[Ecmsc.LIMITS.value][Ecmsc.LIMITS_AMOUNT.value]
         limit_cost = fixed_market_status[Ecmsc.LIMITS.value][Ecmsc.LIMITS_COST.value]
