+++ setup.py
@@ -81,13 +81,14 @@ packages_list = ["octobot_trading.util.initializable",
                  "octobot_trading.data_manager.recent_trades_manager",
                  "octobot_trading.orders.buy_limit_order",
                  "octobot_trading.orders.buy_market_order",
+                 "octobot_trading.orders.order_adapter",
+                 "octobot_trading.orders.order_util",
                  "octobot_trading.orders.sell_limit_order",
                  "octobot_trading.orders.sell_market_order",
                  "octobot_trading.orders.stop_loss_limit_order",
                  "octobot_trading.orders.trailing_stop_order",
                  "octobot_trading.orders.stop_loss_order",
                  "octobot_trading.modes.abstract_trading_mode",
-                 "octobot_trading.modes.trading_mode_orders",
                  "octobot_trading.traders.trader",
                  "octobot_trading.traders.trader_simulator",
                  "octobot_trading.exchanges.exchange_manager",
